Fix build errors and add Swagger package
- Add Swashbuckle.AspNetCore for Swagger/OpenAPI support
- Fix type conversion in RateLimitingService (long to int)
- Remove unused exception variables in RequestAggregationService
- Remove WithOpenApi() calls for NET 10 compatibility
- Project now builds successfully

+++ Services/RateLimitingService.cs
@@ -73,7 +73,7 @@ public class RateLimitingService
             {
                 Limit = policy.RequestsPerMinute,
                 Remaining = policy.RequestsPerMinute,
-                Reset = (long)TimeSpan.FromMinutes(1).TotalSeconds
+                Reset = (int)TimeSpan.FromMinutes(1).TotalSeconds
             };
         }
 
@@ -87,7 +87,7 @@ public class RateLimitingService
         {
             Limit = policy.RequestsPerMinute,
             Remaining = remaining,
-            Reset = entry.GetMinuteWindowSecondsRemaining()
+            Reset = (int)entry.GetMinuteWindowSecondsRemaining()
         };
     }
